+++ data-raw/getSysData.R
@@ -17,6 +17,19 @@ names(evTables) <- tolower(res)
 
 
 # and get the sacramento year type data too
-sacYT <- parseSACData("data-raw/sacrementoData.txt", 1906:2018)
+sacYT <- parseSACData("data-raw/sacrementoData.txt", 1906:2019)
   
-usethis::use_data(evTables, sacYT, internal = TRUE, overwrite = TRUE)
+
+
+# sacramento 4 gage paleo data ----------------
+sac_paleo <- read.table("data-raw/sacramentofourupdate.txt", skip = 3, nrows = 1006)
+sac_paleo2 <- read.table("data-raw/sacramentofourupdate.txt", skip = 1009)
+
+sac_paleo <- rbind(sac_paleo, sac_paleo2[, 1:2])
+sac_paleo$V1 <- zoo::as.yearmon(paste("Dec", sac_paleo$V1))
+
+sac_paleo_wy_vol <- xts::xts(sac_paleo$V2, sac_paleo$V1)
+colnames(sac_paleo_wy_vol) <- "WY_vol"
+
+usethis::use_data(evTables, sacYT, sac_paleo_wy_vol, 
+                  internal = TRUE, overwrite = TRUE)